@@ -1,57 +1,44 @@
 {
-  "name": "contacts_management_system",
-  "version": "1.0.0",
-  "main": "app.js",
-  "scripts": {
-    "build": "webpack",
-    "start": "npm-run-all --parallel open:server lint:watch",
-    "open:server": "webpack-dev-server --open",
-    "lint": "node_modules/.bin/esw src",
-    "lint:watch": "node_modules/.bin/esw src --watch",
-    "fixall": "node_modules/.bin/eslint src --fix"
-  },
-  "keywords": [],
-  "author": "Saket Joshi",
-  "license": "Cybage Software",
+  "name": "my-app",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
     "axios": "^0.16.2",
     "bootstrap": "^3.3.7",
     "query-string": "^5.0.0",
-    "react": "^15.6.1",
-    "react-bootstrap": "^0.31.2",
-    "react-dom": "^15.6.1",
-    "react-favicon": "0.0.9",
-    "react-router-dom": "^4.1.2",
-    "react-table": "^6.5.3",
-    "sha256": "^0.2.0"
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-router": "^6.2.1",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "5.0.0",
+    "sha256": "^0.2.0",
+    "web-vitals": "^2.1.3"
   },
-  "devDependencies": {
-    "babel-core": "^6.25.0",
-    "babel-loader": "^7.1.1",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "css-loader": "^0.28.4",
-    "eslint": "^4.4.1",
-    "eslint-loader": "^1.9.0",
-    "eslint-plugin-react": "^7.2.0",
-    "eslint-watch": "^3.1.2",
-    "extract-text-webpack-plugin": "^2.1.2",
-    "file-loader": "^0.11.2",
-    "html-webpack-plugin": "^2.30.1",
-    "jasmine-core": "^2.7.0",
-    "json-loader": "^0.5.7",
-    "karma": "^1.7.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-jasmine": "^1.1.0",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^2.0.4",
-    "npm-run-all": "^4.0.2",
-    "react-addons-test-utils": "^15.6.0",
-    "react-test-renderer": "^15.6.1",
-    "style-loader": "^0.18.2",
-    "url-loader": "^0.5.9",
-    "webpack": "^3.5.3",
-    "webpack-dev-server": "^2.7.1"
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
