@@ -7,14 +7,15 @@ import React from 'react';
 class Footer extends React.Component {
 	render() {
 		return (
-			<div className="footer navbar-fixed-bottom">
+			// <div className="footer">
 
-				<div className="col-md-12">
-					<h6 className="text-center">
-            Legal | Privacy © 2017 <span className="cyb-footer"> Cybage Software Pvt. Ltd. </span> All Rights Reserved.
-					</h6>
-				</div>
-			</div>
+			// 	<div className="col-md-12">
+			// 		<h6 className="text-center">
+            // Legal | Privacy © 2017 <span className="cyb-footer"> Cybage Software Pvt. Ltd. </span> All Rights Reserved.
+			// 		</h6>
+			// 	</div>
+			// </div>
+			<div />
 		);
 	}
 }
