@@ -6,5 +6,5 @@ import './styles/styles.css';
 
 ReactDOM.render(
 	<App />,
-	document.getElementById('app')
+	document.getElementById('root')
 );
